+++ js/data.js
@@ -1,0 +1,65 @@
+const getDataArrayPosts = () => {
+  const DESCRIPTIONS = [
+    'Вид на пляж с лежаками и отелями.',
+    'Деревянный указатель Идите на пляж.',
+    'Пляж с белым песком и бирюзовым морем.',
+    'Женщина в купальнике на пляже с фотоаппаратом.',
+    'Рис с овощами и мясом в виде смайликов.',
+    'Черный спортивный автомобиль с открытыми дверями.',
+    'Две половинки клубники на деревянной тарелке.',
+    'Красный напиток в стаканах.',
+    'Женщина в купальнике на пляже.',
+    'Система хранения обуви.',
+    'Дорога на пляж.',
+    'Белая Audi.',
+    'Блюдо с красной рыбой и овощами.',
+    'Суши-котик.',
+    'Отдыхаем.',
+    'Вид с высоты птичьего полета.',
+    'Хор в черной одежде.',
+    'Красный классический ретро автомобиль.',
+    'Ноу-хау тапочки с подсветкой.',
+    'Пальмы во дворе на фоне неба.',
+    'Вкуснятина!',
+    'Закат на море.',
+    'Крабик.',
+    'Еее роцк!..!',
+    'Захватывающее автомобильное сафари.',
+  ];
+
+  const COMMENTS = [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+  ];
+
+  const NAMES = [
+    'Степан',
+    'Илья',
+    'Ольга',
+    'Адам',
+    'Полина',
+    'Алексей',
+    'Никита',
+    'Екатерина',
+    'Анна',
+    'Виктория',
+    'Ярослав',
+    'Елисей',
+    'Михаил',
+    'Алиса',
+    'Александра',
+    'Юрий',
+    'Варвара',
+    'Богдан',
+    'Иван',
+    'Агата',
+  ];
+
+  return { DESCRIPTIONS, COMMENTS, NAMES };
+};
+
+export { getDataArrayPosts };
